Update GitHub Actions to use artifact v4 instead of deprecated v3
This PR updates the GitHub workflow files to use `actions/upload-artifact@v4` and `actions/download-artifact@v4` instead of the deprecated v3 versions. This resolves the issue where GitHub Actions are failing with the error:

"This request has been automatically failed because it uses a deprecated version of `actions/upload-artifact: v3`. Learn more: https://github.blog/changelog/2024-04-16-deprecation-notice-v3-of-the-artifact-actions/"

Closes # 13